+++ pauli_tracker/CHANGELOG.md
@@ -18,8 +18,15 @@ methods.
 
 ## [Unreleased]
 ### Added
-- Add the `Tracker::remove_*` methods (with a panicking default implementation to not
-break the trait).
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.4.1] - 2024-05-07
+### Added
+- Add the `Tracker::remove_*` methods (with a panicking default implementation).
 - Add the `Tracker::zc*` methods.
 ### Changed
 ### Deprecated
@@ -202,7 +209,8 @@ them, but hopefully the most important ones.
 
 ## [0.1.0] - 2023-06-01
 
-[Unreleased]: https://github.com/taeruh/pauli_tracker/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/taeruh/pauli_tracker/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/taeruh/pauli_tracker/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/taeruh/pauli_tracker/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/taeruh/pauli_tracker/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/taeruh/pauli_tracker/compare/v0.3.0...v0.3.1